main_svc v.03

+++ main-service/src/main/java/ru/practicum/event/dto/UpdateEventAdminRequestDto.java
@@ -2,9 +2,9 @@ package ru.practicum.event.dto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.AllArgsConstructor;
+import lombok.NoArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
-import lombok.NoArgsConstructor;
 import ru.practicum.enums.EventStateAction;
 import ru.practicum.location.dto.LocationDto;
 import ru.practicum.util.ConvertDataTime;
@@ -30,7 +30,7 @@ public class UpdateEventAdminRequestDto {
     private LocationDto location;
     private Boolean paid;
     @PositiveOrZero
-    private Long participantLimit;
+    private Integer participantLimit;
     private Boolean requestModeration;
     private EventStateAction stateAction;
     @Size(min = 3, max = 120)
